feat(api): 修复扫描统计接口平均健康分数类型不兼容问题
- 修改 `avg_health_score` 字段返回值类型，将 `round()` 返回的 `float` 显式转换为 `float`，确保类型一致性
- 避免因 `round()` 在某些环境下返回 `int` 导致的类型不一致问题

+++ backend/api/v1/scans.py
@@ -103,7 +103,7 @@ async def scan_stats(
     return success_response(data={
         "total": sum(by_status.values()),
         "by_status": by_status,
-        "avg_health_score": round(avg_score, 1) if avg_score else None,
+        "avg_health_score": float(round(avg_score, 1)) if avg_score else None,
     })
